feat: consolidate zoom/pan functionality in ImageViewer and Viewer components; replace deprecated options with unified tweenOptions
Unifying the tween options prevents jitter and sometimes incorrect transformations when tween durations are not the same.

Moving to the `panAndZoom` action instead of multiple separate actions prevents actions fighting each other over the control of the stores. Also proper pinch-to-zoom and mouse wheel zoom was not feasible with separate actions.